chore: tools cleanup & file renames
